adicionado condição para so criar um arquivo env caso ele ainda não exista

+++ backend/docker-entrypoint.sh
@@ -7,11 +7,14 @@ CONTAINER_FIRST_STARTUP="$HOME/.CONTAINER_FIRST_STARTUP"
 if [ ! -f $CONTAINER_FIRST_STARTUP ]; then
   touch $CONTAINER_FIRST_STARTUP
 
-  cat .env.example > .env
-  source .env
+  if [ ! -f .env ]; then
+    cat .env.example > .env
+  fi
 
   cargo install diesel_cli --no-default-features --features postgres
+  source .env
+  diesel migration run --database-url $DATABASE_URL
   cargo build
 fi
 
- cargo run
+cargo run
